Moved the shader related code, and source code, to a separate file.
 * Sorted the sources in CMakeLists
